+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # Unreleased
 
+[...]
+
+# v40.4.0 (21/09/2020)
+
 - **[UPDATE]** Add ability to update `ItemRadioGroup` value programatically
 - **[UPDATE]** Migrate `TimePicker` to new mdx story format
 - **[UPDATE]** Migrate `SelectField` to new mdx story format
 - **[UPDATE]** Migrate `SearchRecap` to new mdx story format
 - **[UPDATE]** Migrate `TopBar` to new mdx story format
 - **[UPDATE]** Migrate `TabsSection` to new mdx story format
-[...]
 
 # v40.3.0 (17/09/2020)